# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: phpcp Control Panel API
  version: "2.0.0"
  description: |
    API ของ Control Panel สำหรับบริหาร PHP Web Hosting

    **การยืนยันตัวตน** ใช้คุกกี้ session (HttpOnly + SameSite=Strict) ที่เซิร์ฟเวอร์ตั้งให้
    ไม่ใช่ bearer token — จงใจไม่ย้าย token ไปเก็บใน JavaScript เพราะจะเป็นการถอยหลัง
    จากสิ่งที่ระบบมีอยู่แล้ว (PLAN-V2 ข้อ N6)

    **CSRF** ทุกคำขอที่เปลี่ยนข้อมูล (POST/PUT/PATCH/DELETE) ต้องส่ง header
    `X-CSRF-Token` ที่ได้จาก `GET /session` · เซิร์ฟเวอร์ส่ง token ตัวใหม่กลับมาทาง
    response header ชื่อเดียวกันทุกครั้งที่มันเปลี่ยน (ตอนล็อกอิน ยืนยัน 2FA
    หรือตอน session id ถูกหมุนตามรอบ) ฝั่ง client ต้องเก็บค่าล่าสุดไว้ใช้เสมอ
servers:
  - url: https://localhost:8443/api/v2
    description: panel ของเครื่องนี้ (พอร์ตเริ่มต้น 8443)
...
paths:
...
  /sites:
    get:
      tags: [sites]
      summary: รายการเว็บไซต์
      description: |
        ผู้ดูแลเว็บไซต์ (webadmin) เห็นเฉพาะเว็บของตัวเอง — กรองที่ระดับ query
        ข้อมูลของลูกค้ารายอื่นจึงไม่ถูกอ่านขึ้นมาตั้งแต่แรก ไม่ใช่กรองทีหลัง
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
        - name: q
          in: query
          schema: { type: string }
          description: ค้นหาจากชื่อเว็บไซต์และโดเมนหลัก
        - name: status
          in: query
          schema: { type: string, enum: [active, suspended] }
        - name: php_version
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema:
            type: string
            enum: [primary_domain, name, created_at, disk_used, php_version, status]
          description: นำหน้าด้วย `-` = จากมากไปน้อย · ชื่อที่ไม่รู้จักตกกลับไปใช้ค่าปริยาย
      responses:
        "200":
          description: รายการเว็บไซต์พร้อม meta ของการแบ่งหน้า
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Site" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }

    post:
      tags: [sites]
      summary: สร้างเว็บไซต์ใหม่
      description: |
        สั่งผ่าน capability `site.create` — สร้าง system user, FPM pool, vhost และ
        โครงสร้างโฟลเดอร์ให้ครบในคำสั่งเดียว · ล้มกลางทางจะคืนค่าทั้งหมดให้เอง

        `owner_user_id` ถูกบังคับเป็น id ของผู้เรียกเสมอเมื่อผู้เรียกเป็น webadmin
        ค่าที่ส่งมาจะถูกเมิน — ลูกค้าสร้างเว็บให้คนอื่นไม่ได้
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain, php_version]
              properties:
                domain: { type: string, maxLength: 253 }
                name: { type: string, description: "ว่าง = ใช้ชื่อโดเมน" }
                php_version: { type: string, example: "8.4" }
                aliases:
                  oneOf:
                    - { type: array, items: { type: string } }
                    - { type: string, description: "คั่นด้วยเว้นวรรคหรือจุลภาค" }
                docroot: { type: string, description: "ว่าง = <บ้าน>/public" }
                pointer_root: { type: string }
                owner_user_id: { type: integer }
      responses:
        "201":
          description: สร้างแล้ว
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data: { $ref: "#/components/schemas/Site" }
        "409":
          description: "`CONFLICT` — โดเมนนี้มีอยู่แล้ว"
          $ref: "#/components/responses/Error"
        "422": { $ref: "#/components/responses/Error" }
        "503":
          description: "`AGENT_UNAVAILABLE` — ติดต่อ phpcp-agentd ไม่ได้ ลองใหม่ได้เมื่อบริการกลับมา"
          $ref: "#/components/responses/Error"
  /sites/{id}:
    parameters:
      - { $ref: "#/components/parameters/SiteId" }
    get:
      tags: [sites]
      summary: รายละเอียดเว็บไซต์พร้อมโดเมนทั้งหมด
      description: |
        `id=0` เป็นค่าจองพิเศษ — แปลว่า "เว็บนี้ยังไม่มีอยู่จริง ขอค่าเริ่มต้นสำหรับสร้างใหม่"
        แทนที่จะเป็น 404 · หน้าฟอร์มสร้างเว็บไซต์ใช้เส้นทางเดียวกับหน้าแก้ไขได้ ไม่ต้องมี
        endpoint แยกต่างหากที่ทำหน้าที่คล้ายกันแค่เพื่อกรณีสร้างใหม่ (ดู `NewSiteDefaults`)
      responses:
        "200":
          description: เว็บไซต์ หรือค่าเริ่มต้นสำหรับสร้างใหม่ (เมื่อ id=0)
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      ok: { type: boolean, const: true }
                      data:
                        allOf:
                          - { $ref: "#/components/schemas/Site" }
                          - type: object
                            properties:
                              domains:
                                type: array
                                items: { $ref: "#/components/schemas/Domain" }
                              aliases: { type: array, items: { type: string } }
                  - type: object
                    properties:
                      ok: { type: boolean, const: true }
                      data: { $ref: "#/components/schemas/NewSiteDefaults" }
        "404":
          description: |
            ไม่พบ **หรือ** ไม่ใช่เว็บของผู้เรียก — ตอบเหมือนกันโดยเจตนา
            ถ้าตอบ 403 เมื่อ "มีอยู่แต่ไม่ใช่ของคุณ" จะใช้ไล่เดา id เพื่อสำรวจเครื่องได้
            (ไม่เกิดกับ id=0 — เคสนั้นตอบ 200 เสมอ)
          $ref: "#/components/responses/Error"

    patch:
      tags: [sites]
      summary: แก้ไขบางส่วน
      description: |
        รองรับเฉพาะ `php_version` เพราะเป็นฟิลด์เดียวของเว็บไซต์ที่มี capability รองรับ
        การเปลี่ยนชื่อหรือ docroot ต้องรอ capability ของมันเอง — ชั้น API จะไม่เขียน
        ฐานข้อมูลตรง ๆ เพราะนั่นคือการข้ามชั้นที่ 2 และทำให้ไม่มีร่องรอยใน audit log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                php_version: { type: string }
      responses:
        "200":
          description: เว็บไซต์หลังแก้ไข
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data: { $ref: "#/components/schemas/Site" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }

    delete:
      tags: [sites]
      summary: ลบเว็บไซต์
      description: ต้องยืนยันด้วยชื่อโดเมนเสมอ · ส่งทาง body หรือ query ก็ได้
      parameters:
        - name: confirm_domain
          in: query
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                confirm_domain: { type: string }
      responses:
        "200":
          description: สำเร็จ — คำตอบของคำสั่งพร้อม `actions` ให้หน้าจอทำตาม
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandResponse" }
        "404": { $ref: "#/components/responses/Error" }
        "422":
          description: "`VALIDATION_ERROR` — ชื่อโดเมนที่ยืนยันไม่ตรง"
          $ref: "#/components/responses/Error"
...
  /sites/{id}/domains:
    parameters:
      - { $ref: "#/components/parameters/SiteId" }
    get:
      tags: [sites]
      summary: โดเมนทั้งหมดของเว็บไซต์นี้
      responses:
        "200":
          description: รายการโดเมน
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Domain" }
        "404": { $ref: "#/components/responses/Error" }

    post:
      tags: [sites]
      summary: เพิ่มโดเมนย่อยหรือ alias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host]
              properties:
                host: { type: string, description: "ชื่อโฮสต์ย่อย หรือโดเมนเต็ม" }
                path: { type: string, description: "โฟลเดอร์ปลายทางของโดเมนย่อย" }
      responses:
        "201":
          description: เพิ่มแล้ว
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data: { $ref: "#/components/schemas/Domain" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }

    put:
      tags: [sites]
      summary: แทนที่รายการโดเมนทั้งชุดตามชนิด
      description: capability คำนวณเองว่าต้องเพิ่มหรือลบอะไรเพื่อให้ตรงกับรายการที่ส่งมา
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domains]
              properties:
                domains:
                  oneOf:
                    - { type: array, items: { type: string } }
                    - { type: string }
                type: { type: string, enum: [alias, subdomain], default: alias }
      responses:
        "200":
          description: ผลการแทนที่
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data: { type: object }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /domains/{id}:
    delete:
      tags: [domains]
      summary: ลบโดเมน
      parameters:
        - { $ref: "#/components/parameters/DomainId" }
      responses:
        "200":
          description: สำเร็จ — คำตอบของคำสั่งพร้อม `actions` ให้หน้าจอทำตาม
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandResponse" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: "`CONFLICT` — เป็นโดเมนหลักของเว็บไซต์ ลบแยกไม่ได้ ต้องลบทั้งเว็บ"
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      description: ค่าที่เกินเพดานถูกบีบลงมาที่ 200 ไม่ใช่ปฏิเสธคำขอ
    SiteId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    DomainId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: phpcp_sid
      description: |
        ตั้งโดยเซิร์ฟเวอร์เท่านั้น · HttpOnly + SameSite=Strict + Secure (เมื่อใช้ HTTPS)
        ชื่อจะเป็น `__Host-phpcp_sid` เมื่อเปิด cookie_secure
  schemas:
...
    PageMeta:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
...
  responses:
    Error:
      description: ข้อผิดพลาดตามรูปแบบมาตรฐาน
      headers:
        X-CSRF-Token:
          schema: { type: string }
          description: แนบมากับ 419 เสมอ เพื่อให้ลองใหม่ได้ทันทีโดยไม่ต้องรีเฟรชหน้า
        Retry-After:
          schema: { type: integer }
          description: แนบมากับ 429 เสมอ
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
security:
  - sessionCookie: []